+++ lib/controllers/rtc_event_handlers.dart
@@ -79,8 +79,20 @@ Future<RtcEngineEventHandler> rtcEngineEventHandler(
     agoraEventHandlers.onFirstRemoteVideoDecoded
         ?.call(connection, remoteUid, width, height, elapsed);
   }, onUserMuteVideo: (connection, remoteUid, muted) {
+    log("onUserMuteVideo- muted: $muted, uid: $remoteUid",
+        name: tag, error: Level.info.value);
+    if (remoteUid != sessionController.value.localUid) {
+      sessionController.updateUserVideo(uid: remoteUid, videoDisabled: muted);
+    }
+
     agoraEventHandlers.onUserMuteVideo?.call(connection, remoteUid, muted);
   }, onUserMuteAudio: (connection, remoteUid, muted) {
+    log("onUserMuteAudio- muted: $muted, uid: $remoteUid",
+        name: tag, error: Level.info.value);
+    if (remoteUid != sessionController.value.localUid) {
+      sessionController.updateUserAudio(uid: remoteUid, muted: muted);
+    }
+
     agoraEventHandlers.onUserMuteAudio?.call(connection, remoteUid, muted);
   }, onVideoSizeChanged: (connection, sourceType, uid, width, height, rotation) {
     agoraEventHandlers.onVideoSizeChanged
@@ -223,7 +235,9 @@ Future<RtcEngineEventHandler> rtcEngineEventHandler(
       if (state == RemoteVideoState.remoteVideoStateStopped) {
         sessionController.updateUserVideo(uid: remoteUid, videoDisabled: true);
       } else if (state == RemoteVideoState.remoteVideoStateDecoding &&
-          reason == RemoteVideoState.remoteVideoStateStarting) {
+          (reason == RemoteVideoState.remoteVideoStateStarting ||
+              reason ==
+                  RemoteVideoStateReason.remoteVideoStateReasonRemoteUnmuted)) {
         sessionController.updateUserVideo(uid: remoteUid, videoDisabled: false);
       }
     }