+++ lib/src/layout/widgets/disabled_video_widget.dart
@@ -15,10 +15,10 @@ class _DisabledVideoWidgetState extends State<DisabledVideoWidget> {
       color: Colors.black,
       padding: const EdgeInsets.all(8.0),
       child: Center(
-        child: Image.network(
-          'https://i.ibb.co/q5RysSV/image.png',
-        ),
-      ),
+          child: Icon(
+        Icons.person_rounded,
+        color: Colors.white,
+      )),
     );
   }
 }